requirements.txt
# =========================
# Core Data Stack
# =========================
pandas>=2.1.0
numpy>=1.26.0
sqlalchemy>=2.0.20

# =========================
# Visualization / Notebooks
# =========================
matplotlib>=3.8.0
seaborn>=0.13.2
jupyter
ipykernel>=6.28.0

# =========================
# Job Scraping
# =========================
python-jobspy>=0.31.0

# =========================
# Utilities
# =========================
tqdm

# =========================
# NLP / Embeddings Core
# =========================
sentence-transformers>=2.7.0
transformers>=4.41.0
torch>=2.2.0

# Supporting NLP dependencies (kept explicit for stability)
regex>=2024.11.6
scipy>=1.11.0
scikit-learn>=1.4.0
huggingface-hub>=0.23.0
tokenizers>=0.19.0
safetensors>=0.4.0

# =========================
# Optional: Skill Extraction
# =========================
skillNer
pyresparser